requirements.txt
beautifulsoup4==4.12.2
lyricsgenius==3.0.1
openai==0.27.2
openai_whisper==20230314
pandas==1.5.3
pinecone_client==2.2.1
python-dotenv==1.0.0
requests==2.28.2
scikit_learn==1.2.2
SpeechRecognition==3.10.0
spotipy==2.23.0
whisper==1.1.10
pyaudio
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git